quickstart-python/requirements.txt
requests==2.11.1
PyJWT==1.4.2
